Order Fixes

+++ NamonaProject_v3/NamonaProject_v3/Model/OrderModel.cs
@@ -13,19 +13,55 @@ namespace NamonaProject_v3_.Model
         {
             _context = context;
         }
+        public MyCartDto MyCart(int userid)
+        {
+            if (!_context.orders.Any(x => x.OrderId == userid))
+            {
+                throw new KeyNotFoundException("nincs ilyen ruha");
+            }
+            var carts = _context.cart
+                    .Where(c => c.UserId == userid ) // csak aktív kosár
+                    .Include(c => c.Clothing).ThenInclude(cl => cl.Category)
+                    .Include(c => c.Clothing).ThenInclude(cl => cl.Gender)
+                     .Select(c => new CartItemDto
+                     {
+                         ClothingId = c.ClothingId,
+                         ClothingName = c.Clothing.ClothingName,
+                         Collection = c.Clothing.Collection,
+                         CategoryId = c.Clothing.CategoryId,
+                         CategoryName = c.Clothing.Category.CategoryName,
+                         Color = c.Clothing.Color,
+                         Price = c.Clothing.Price,
+                         PriceSum = c.PriceSum,
+                         Stock = c.Clothing.Stock,
+                         Amount = c.Amount,
+                         GenderId = c.Clothing.GenderId,
+                         GenderName = c.Clothing.Gender.GenderType
+                     }).ToList();
+
+            return new MyCartDto
+            {
+                UserId = userid,
+                Carts = carts
+            };
+        }
 
         public IEnumerable<OrderDto> AllOrders()
         {
             return _context.orders.Select(x => new OrderDto
             {
                 OrderId = x.OrderId,
-                OrderDate = x.OrderDate,
+                OrderDate = (DateTimeOffset)x.OrderDate,
                 Address = x.Address
             });
         }
 
         public async Task DeleteOrder(int id)
         {
+            if (!_context.orders.Any(x => x.OrderId == id))
+            {
+                throw new KeyNotFoundException("nincs ilyen ruha");
+            }
             using (var trx = _context.Database.BeginTransaction())
             {
                 _context.orders.Remove(_context.orders.Where(x => x.OrderId == id).First());
@@ -35,14 +71,16 @@ namespace NamonaProject_v3_.Model
             await Task.CompletedTask;
         }
 
-        public async Task AddOrder(OrderDto order)
+        public async Task AddOrder(AddOrderDto order)
         {
             using (var trx = _context.Database.BeginTransaction())
             {
                 _context.orders.Add(new Persistance.Orders
                 {
                     OrderDate = order.OrderDate,
-                    Address = order.Address
+                    Address = order.Address,
+                    CompletedAt = null
+
                 });
                 await _context.SaveChangesAsync();
                 await trx.CommitAsync();
@@ -50,13 +88,17 @@ namespace NamonaProject_v3_.Model
             await Task.CompletedTask;
         }
 
-        public async Task UpdateOrder(int id, OrderDto order)
+        public async Task UpdateOrder(OrderDto order)
         {
+            if (!_context.orders.Any(x => x.OrderId == order.OrderId))
+            {
+                throw new KeyNotFoundException("nincs ilyen ruha");
+            }
+            var existingOrder = _context.orders.Where(x => x.OrderId == order.OrderId).First();
             using (var trx = _context.Database.BeginTransaction())
             {
-                var existingOrder = _context.orders.Where(x => x.OrderId == id).First();
-                existingOrder.OrderDate = order.OrderDate;
                 existingOrder.Address = order.Address;
+                existingOrder.Status = order.Status;
                 await _context.SaveChangesAsync();
                 await trx.CommitAsync();
             }
@@ -70,13 +112,17 @@ namespace NamonaProject_v3_.Model
             var order = _context.orders.FirstOrDefault(x => x.OrderId == id);
 
             if (order == null)
-                throw new Exception("Order not found");
+                throw new KeyNotFoundException("Order not found");
+            using (var trx = _context.Database.BeginTransaction())
+            {
+                order.Status = "Completed";
+                order.CompletedAt = DateTime.UtcNow;
 
-            order.Status = "Completed";
-            order.CompletedAt = DateTime.UtcNow;
 
-            await _context.SaveChangesAsync();
-          //  await trx.CommitAsync();
+                await _context.SaveChangesAsync();
+                await trx.CommitAsync();
+            }
+            await Task.CompletedTask;
         }
 
 
@@ -91,19 +137,22 @@ namespace NamonaProject_v3_.Model
             await Task.CompletedTask;
         }
 
-        public async Task UpdateOrderStatus(int id, string status)
+        public async Task UpdateOrderStatus(UpdateStatusDto dto)
         {
-            var order = _context.orders.FirstOrDefault(x => x.OrderId == id);
+            var order = _context.orders.FirstOrDefault(x => x.OrderId == dto.OrderId);
             if (order == null)
-                throw new Exception("Order not found");
-            order.Status = status;
-            await _context.SaveChangesAsync();
-            //await trx.CommitAsync();
+                throw new KeyNotFoundException("Order not found");
+            using (var trx = _context.Database.BeginTransaction())
+            {
+                order.Status = dto.Status;
+                await _context.SaveChangesAsync();
+                await trx.CommitAsync();
 
+            }
             await Task.CompletedTask;
         }
 
-        
+
 
     }
 }
